Added notion of locked state, and optional link. Converted JS to use class.
git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@102106 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/web/eiffel-org/modules/codeboard/codeboard_module_webapi.e
@@ -23,9 +23,11 @@ feature -- Security
 		do
 			Result := Precursor
 			Result.force (manage_snippet_permission)
+			Result.force (edit_snippet_permission)
 		end
 
 	manage_snippet_permission: STRING = "manage codeboard snippet"
+	edit_snippet_permission: STRING = "edit codeboard snippet"
 
 feature -- Router/webapi
 
@@ -64,8 +66,15 @@ feature -- Handlers
 					else
 						r.add_string_field ("lang", "eiffel") -- Default
 					end
-					if attached l_code.description as l_desc then
-						r.add_string_field ("description", l_desc)
+
+					if l_code.is_locked then
+						r.add_boolean_field ("is_locked", True)
+					end
+					if attached l_code.link as l_link then
+						r.add_string_field ("link", l_link)
+					end
+					if attached l_code.caption as l_caption then
+						r.add_string_field ("caption", l_caption)
 					end
 					r.add_string_field ("code", l_code.text)
 					if l_code_id = l_code_count then
@@ -81,7 +90,7 @@ feature -- Handlers
 					r.add_link ("next", "code#" + (l_code_id + 1).out, a_codeboard_api.cms_api.webapi_path (codeboard_code_location + "1"))
 				end
 			else
-				r.add_integer_64_field ("available_code_count", l_code_count)
+				r.add_integer_64_field ("code_count", l_code_count)
 				if l_code_count > 0 then
 					from
 						i := 1
@@ -105,9 +114,10 @@ feature -- Handlers
 			l_code_id: INTEGER
 			l_json_text: STRING
 			utf: UTF_CONVERTER
+			l_change_allowed, l_perm_denied: BOOLEAN
 		do
 			create resp.make (req, res, a_codeboard_api.cms_api)
-			if a_codeboard_api.cms_api.has_permission (manage_snippet_permission) then
+			if a_codeboard_api.cms_api.has_permissions (<<manage_snippet_permission, edit_snippet_permission>>) then
 
 				if req.is_put_post_request_method or req.is_delete_request_method then
 					if
@@ -126,10 +136,18 @@ feature -- Handlers
 							resp.set_status_code ({HTTP_STATUS_CODE}.method_not_allowed)
 							resp.add_string_field ("error", "POST method not allowed!")
 						elseif req.is_delete_request_method then
-							a_codeboard_api.delete_code (l_code_id)
-							if a_codeboard_api.has_error then
-								resp.set_status_code ({HTTP_STATUS_CODE}.internal_server_error)
-								resp.add_string_field ("error", "Deletion failed!")
+							if a_codeboard_api.is_code_locked (l_code_id) then
+								resp.set_status_code ({HTTP_STATUS_CODE}.precondition_failed)
+								resp.add_string_field ("error", "Code " + l_code_id.out + " is locked!")
+							elseif a_codeboard_api.cms_api.has_permission (manage_snippet_permission) then
+								a_codeboard_api.delete_code (l_code_id)
+								if a_codeboard_api.has_error then
+									resp.set_status_code ({HTTP_STATUS_CODE}.internal_server_error)
+									resp.add_string_field ("error", "Deletion failed!")
+								end
+							else
+								resp.set_status_code ({HTTP_STATUS_CODE}.user_access_denied)
+								resp.add_string_field ("error", "Permission denied!")
 							end
 						elseif req.is_put_post_request_method then
 							if attached req.content_type as ct and then ct.same_string ("application/json") then
@@ -138,12 +156,39 @@ feature -- Handlers
 								if req.has_error then
 									resp.set_status_code ({HTTP_STATUS_CODE}.bad_request)
 									resp.add_string_field ("error", "Missing data!")
-								elseif attached json_to_string (utf.utf_8_string_8_to_string_32 (l_json_text)) as l_snippet then
-									l_snippet.set_id (l_code_id)
-									a_codeboard_api.update_code (l_snippet)
-									if a_codeboard_api.has_error then
-										resp.set_status_code ({HTTP_STATUS_CODE}.internal_server_error)
-										resp.add_string_field ("error", "Update failed!")
+								elseif attached json_to_snippet (utf.utf_8_string_8_to_string_32 (l_json_text)) as l_snippet then
+									if a_codeboard_api.is_code_locked (l_code_id) then
+											-- Check if this is a request to unlock the code.
+										if
+											not l_snippet.is_locked and then
+											attached a_codeboard_api.code (l_code_id) as l_locked_code
+										then
+											if a_codeboard_api.cms_api.has_permission (manage_snippet_permission) then
+													-- Check if `is_locked` is the only difference!
+												l_locked_code.unlock
+												l_change_allowed := l_locked_code.same_as (l_snippet)
+												l_locked_code.lock -- restore previous state.
+											else
+												l_change_allowed := False
+												l_perm_denied := True
+											end
+										end
+									else
+										l_change_allowed := True
+									end
+									if l_change_allowed then
+										l_snippet.set_id (l_code_id)
+										a_codeboard_api.update_code (l_snippet)
+										if a_codeboard_api.has_error then
+											resp.set_status_code ({HTTP_STATUS_CODE}.internal_server_error)
+											resp.add_string_field ("error", "Update failed!")
+										end
+									elseif l_perm_denied then
+										resp.set_status_code ({HTTP_STATUS_CODE}.user_access_denied)
+										resp.add_string_field ("error", "Could not unlock code (permission denied)!")
+									else
+										resp.set_status_code ({HTTP_STATUS_CODE}.precondition_failed)
+										resp.add_string_field ("error", "Code " + l_code_id.out + " is locked!")
 									end
 								else
 									resp.set_status_code ({HTTP_STATUS_CODE}.bad_request)
@@ -163,7 +208,7 @@ feature -- Handlers
 								if req.has_error then
 									resp.set_status_code ({HTTP_STATUS_CODE}.bad_request)
 									resp.add_string_field ("error", "Missing data!")
-								elseif attached json_to_string (utf.utf_8_string_8_to_string_32 (l_json_text)) as l_snippet then
+								elseif attached json_to_snippet (utf.utf_8_string_8_to_string_32 (l_json_text)) as l_snippet then
 									if l_snippet.has_id then
 										a_codeboard_api.insert_code (l_snippet.id, l_snippet)
 									else
@@ -203,7 +248,7 @@ feature -- Handlers
 			resp.execute
 		end
 
-	json_to_string (a_json_text: READABLE_STRING_GENERAL): detachable CODEBOARD_SNIPPET
+	json_to_snippet (a_json_text: READABLE_STRING_GENERAL): detachable CODEBOARD_SNIPPET
 		local
 			jp: JSON_PARSER
 		do
@@ -217,11 +262,22 @@ feature -- Handlers
 					elseif attached {JSON_STRING} jo.item ("code_id") as s_code_id then
 						Result.set_id (s_code_id.unescaped_string_32.to_integer_32)
 					end
-					if attached {JSON_STRING} jo.item ("description") as l_desc then
-						Result.set_description (l_desc.unescaped_string_32)
+					if attached {JSON_STRING} jo.item ("caption") as l_caption then
+						Result.set_caption (l_caption.unescaped_string_32)
 					end
 					if attached {JSON_STRING} jo.item ("lang") as l_lang then
 						Result.set_lang (l_lang.unescaped_string_8)
+					end
+					if attached {JSON_BOOLEAN} jo.item ("is_locked") as l_is_locked then
+						Result.set_is_locked (l_is_locked.item)
+					end
+					if attached {JSON_STRING} jo.item ("link") as js_link then
+						Result.set_link (js_link.unescaped_string_8)
+					elseif
+						attached {JSON_OBJECT} jo.item ("link") as j_link and then
+						attached {JSON_STRING} j_link.item ("href") as j_link_href
+					then
+						Result.set_link (j_link_href.unescaped_string_8)
 					end
 				end
 			end
